✨ feat(ast): 修改行号类型

+++ neige_compiler/src/ast/exp.rs
@@ -2,49 +2,49 @@ use super::Block;
 
 #[derive(Debug, PartialEq)]
 pub struct StringExp {
-    line: i32,
+    line: usize,
     str: String,
 }
 
 #[derive(Debug, PartialEq)]
 pub struct NameExp {
-    line: i32,
+    line: usize,
     name: String,
 }
 
 #[derive(Debug, PartialEq)]
 pub struct UnopExp {
-    line: i32,
-    op: i32,
+    line: usize,
+    op: usize,
     exp: Box<Exp>,
 }
 
 #[derive(Debug, PartialEq)]
 pub struct BinopExp {
-    line: i32,
-    op: i32,
+    line: usize,
+    op: usize,
     exp1: Box<Exp>,
     exp2: Box<Exp>,
 }
 
 #[derive(Debug, PartialEq)]
 pub struct ConcatExp {
-    line: i32,
+    line: usize,
     exps: Vec<Box<Exp>>,
 }
 
 #[derive(Debug, PartialEq)]
 pub struct TableConstructorExp {
-    line: i32,
-    last_line: i32,
+    line: usize,
+    last_line: usize,
     key_exps: Vec<Box<Exp>>,
     val_exps: Vec<Box<Exp>>,
 }
 
 #[derive(Debug, PartialEq)]
 pub struct FuncCallExp {
-    line: i32,
-    last_line: i32,
+    line: usize,
+    last_line: usize,
     prefix_exp: Box<Exp>,
     name_exp: Option<Box<StringExp>>,
     args: Vec<Box<Exp>>,
@@ -52,8 +52,8 @@ pub struct FuncCallExp {
 
 #[derive(Debug, PartialEq)]
 pub struct FuncDefExp {
-    line: i32,
-    last_line: i32,
+    line: usize,
+    last_line: usize,
     par_list: Vec<String>,
     is_vararg: bool,
     block: Box<Block>,
@@ -61,7 +61,7 @@ pub struct FuncDefExp {
 
 #[derive(Debug, PartialEq)]
 pub struct TableAccessExp {
-    last_line: i32,
+    last_line: usize,
     prefix_exp: Box<Exp>,
     key_exp: Box<Exp>,
 }
@@ -69,12 +69,12 @@ pub struct TableAccessExp {
 #[allow(dead_code)]
 #[derive(Debug, PartialEq)]
 pub enum Exp {
-    Nil(i32),
-    True(i32),
-    False(i32),
-    Vararg(i32),
-    Integer(i32, i64),
-    Float(i32, f64),
+    Nil(usize),
+    True(usize),
+    False(usize),
+    Vararg(usize),
+    Integer(usize, i64),
+    Float(usize, f64),
     String(StringExp),
     Name(NameExp),
     Unop(UnopExp),
