+++ neige_compiler/src/ast/stat.rs
@@ -23,8 +23,8 @@ pub struct IfStat {
 
 #[derive(Debug, PartialEq)]
 pub struct ForNumStat {
-    line_of_for: i32,
-    line_of_do: i32,
+    line_of_for: usize,
+    line_of_do: usize,
     var_name: String,
     init_exp: Box<Exp>,
     limit_exp: Box<Exp>,
@@ -34,7 +34,7 @@ pub struct ForNumStat {
 
 #[derive(Debug, PartialEq)]
 pub struct ForInStat {
-    line_of_do: i32,
+    line_of_do: usize,
     name_list: Vec<String>,
     exp_list: Vec<Box<Exp>>,
     block: Box<Block>,
@@ -42,14 +42,14 @@ pub struct ForInStat {
 
 #[derive(Debug, PartialEq)]
 pub struct LocalVarDeclStat {
-    last_line: i32,
+    last_line: usize,
     name_list: Vec<String>,
     exp_list: Vec<Box<Exp>>,
 }
 
 #[derive(Debug, PartialEq)]
 pub struct AssignStat {
-    last_line: i32,
+    last_line: usize,
     var_list: Vec<Box<Exp>>,
     exp_list: Vec<Box<Exp>>,
 }
@@ -63,7 +63,7 @@ pub struct LocalFuncDefStat {
 #[allow(dead_code)]
 #[derive(Debug, PartialEq)]
 pub enum Stat {
-    Break(i32),
+    Break(usize),
     Label(String),
     Goto(String),
     Do(Box<Block>),
